added clgs

+++ scripts/downloadHandler.js
@@ -4,26 +4,30 @@ const deleteFolderRecursive = require('./downloadHandler_deleteFolder.js');
 
 const source = process.argv[2];
 const destination = process.argv[3];
-const deleteFilesOlderThanDays = Number(process.argv[4]);
-const fileTypes = process.argv.slice(5).map(filetype => `.${filetype}`);
+const keepSourceFolder = process.argv[4].toUpperCase();
+let fileTypes;
 
-console.log('Copying only filetypes: ' , fileTypes);
+// set fileTypes
 
-// copia todos os que estejam dentro do intervalo de datas
-// copia só os tipos de ficheiros que lhe passo
+if (process.argv[5]) {
+  fileTypes = process.argv.slice(5).map(filetype => `.${filetype}`);
+} else {
+  fileTypes = '*';
+}
+
 // progress bar
 // show usage and wait for input
 
-if (validInput(source, destination, deleteFilesOlderThanDays)) {
-  copyFolderRecursive.source = source;
+if (validInput(source, destination)) {
   deleteFolderRecursive.rootPath = source;
-  copyFolderRecursive(source,destination,fileTypes);
-  deleteFolderRecursive(source);
+  copyFolderRecursive(source, destination, fileTypes);
+  if (keepSourceFolder === 'N')deleteFolderRecursive(source);
+  console.log('All done!');
 } else {
   process.exit();
 }
 
-function validInput(source, destination, deleteFilesOlderThanDays) {
+function validInput(source, destination) {
   if (!fs.existsSync(source)) {
     process.on('exit', () => {
       console.log('Oops! Files not copied. Source path is invalid.');
@@ -34,10 +38,10 @@ function validInput(source, destination, deleteFilesOlderThanDays) {
       console.log('Oops! Files not copied. Destination path is invalid.');
     });
     return false;
-  } else if (deleteFilesOlderThanDays && typeof deleteFilesOlderThanDays !== 'number' && !isNaN(deleteFilesOlderThanDays)) {
-    process.on('exit', () => {
-      console.log('Oops! Files not copied. Please input a valid number.');
-    });
+  } else if (!(keepSourceFolder === 'Y' || keepSourceFolder === 'N')) {
+    process.on('exit' , () => {
+        console.log('Please input a valid string (y/n) to signal if you want to keep source folder.');
+    })
     return false;
   }
   return true;
